Cosmetics.
git-svn-id: https://nemerle.googlecode.com/svn/nemerle/trunk@5812 c8a6711f-211a-0410-a8d5-2f220496d6d1

+++ Makefile
@@ -75,18 +75,20 @@ send: dist send-dist
 
 changelog:
 	svn up
-	last=`perl -e '$$_ = <>; /\[r(\d+-)?(\d+)\]/ and print "$$2\n"' ChangeLog 2>/dev/null`; \
+	@echo Logging...
+	@last=`perl -e '$$_ = <>; /\[r(\d+-)?(\d+)\]/ and print "$$2\n"' ChangeLog 2>/dev/null`; \
 	if [ X$$last = X ] ; then last=0 ; fi ; \
 	last=$$(($$last + 1)) ; \
 	if [ `svn info | grep '^Revision:' | sed -e 's/.* //'` -le $$last ] ; then \
-         echo "Nothing to log." ; \
+           echo "Nothing to log." ; \
        	else \
-         echo "Loggin from $$last."; \
-         svn log -r HEAD:$$last -v --xml > changelog.xml  && \
-         $(svn2log) -o ChangeLog.new -p '$(nemroot)' -r npc/ncc=ncc && \
-         cat ChangeLog > ChangeLog.old 2>/dev/null; \
-         cat ChangeLog.new ChangeLog.old > ChangeLog && \
-         rm -f ChangeLog.old ChangeLog.new changelog.xml ; \
+           echo "Loggin from $$last."; \
+           svn log -r HEAD:$$last -v --xml > changelog.xml  && \
+	   echo Running the script. && \
+           $(svn2log) -o ChangeLog.new -p '$(nemroot)' -r npc/ncc=ncc && \
+           cat ChangeLog > ChangeLog.old 2>/dev/null; \
+           cat ChangeLog.new ChangeLog.old > ChangeLog && \
+           rm -f ChangeLog.old ChangeLog.new changelog.xml ; \
         fi
 
 sync-boot:
